nit: small cleaning
